# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: VITAL Platform API
  description: |
    AI-powered healthcare intelligence platform API.
    
    ## Authentication
    
    All endpoints require Bearer token authentication via Supabase Auth.
    Include the token in the `Authorization` header:
    
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Multi-tenancy
    
    All data is isolated by `organization_id`. The organization context is 
    automatically extracted from the JWT token.
    
    ## Rate Limiting
    
    - Token budget enforced per organization
    - 429 returned when budget exceeded
    - Check `/health` for current usage
  version: 3.0.0
  contact:
    name: VITAL Platform Team
    email: platform@vital.ai
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.vital.ai
    description: Production
...
paths:
...
  /api/jobs/{id}/status:
    get:
      tags: [Jobs]
      summary: Get job status
      operationId: getJobStatus
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
        '404':
          description: Job not found
  /api/jobs/{id}/result:
    get:
      tags: [Jobs]
      summary: Get job result
      operationId: getJobResult
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResult'
        '404':
          description: Job not found
  /api/jobs/{id}/cancel:
    post:
      tags: [Jobs]
      summary: Cancel job
      operationId: cancelJob
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job cancelled
        '404':
          description: Job not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    JobStatus:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled]
        progress:
          type: number
          minimum: 0
          maximum: 100
        message:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    JobResult:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
        result:
          type: object
        error:
          type: string
...